More expanduser
Eg, `Path("~/cirrus").resolve()` will expand to `Path.cwd() /
"~/cirrus"`, which is not expected. This `.expanduser()` to a few
`Path`s where the user may reasonably expect `~` to work as replacement
for their home directory. (Note: In a lot of cases, this will have been
done by the user's shell)